feat:(Optimization-#51): Optimization of some features (GlobalManager, Helpers..)
