+++ app/Models/Comments/CommentsManager.php
@@ -1,14 +1,12 @@
 <?php
-namespace App\Models;
+namespace App\Models\Comments;
 
-use App\Core\Database;
-use PDOStatement;
+use App\Models\Articles\ArticlesModel;
+use App\Models\GlobalManager;
+use App\Models\Users\UsersModel;
 
-class CommentsManager extends Database
+class CommentsManager extends GlobalManager
 {
-
-    private Database $db;
-
 
     public function countAllInvalid()
     {
@@ -65,7 +63,12 @@ class CommentsManager extends Database
         return $data;
     }
 
-    public function findAllValid($id)
+    public function countAllValidFromArticle(int $id)
+    {
+        return $this->request("SELECT COUNT(*) as nb_comments FROM comments WHERE article_id = :id AND validate = 1", ['id' => $id])->fetch();
+    }
+
+    public function findAllValid($id, int $limit, int $perPage)
     {
         // Query
         $sql = "SELECT 
@@ -83,7 +86,8 @@ class CommentsManager extends Database
                 INNER JOIN users as B on A.author_id = B.id
                 INNER JOIN users as C on A.validate_by = C.id
                 WHERE article_id = :id AND A.validate = 1 
-                ORDER BY created_at DESC";
+                ORDER BY created_at DESC
+                LIMIT $limit, $perPage";
 
         // Execute Request
         $results = $this->request($sql, ['id' => $id])->fetchAll();
@@ -114,81 +118,11 @@ class CommentsManager extends Database
         return $data;
     }
 
-    public function findAllBy(string $params, string $value)
-    {
-        // Query 
-        $sql = "SELECT * FROM comments WHERE $params = :value";
-
-        // Execute request
-        $result = $this->request($sql, ['value' => $value])->fetchAll();
-
-        return $result;
-    }
-
     public function validComment($comment_id, $admin_id) 
     {
         $sql = "UPDATE comments SET validate = 1, validate_by = :admin_id WHERE id = :comment_id";
 
         return $this->request($sql, ['admin_id' => $admin_id, 'comment_id' => $comment_id]);
-    }
-
-    public function create()
-    {
-        $keys = [];
-        $inter = [];
-        $values = [];
-
-        // Loop to get parameters and values and add inter("?")
-        foreach ($this as $key => $value) {
-            if ($value !== null && $key != 'db' && $key != 'table') {
-                $keys[] = $key;
-                $inter[] = "?";
-                $values[] = $value;
-            }
-        }
-
-        // Transforms array into a string
-        $list_keys = implode(', ', $keys);
-        $list_inter = implode(', ', $inter);
-
-        // Execute request
-        return $this->request('INSERT INTO comments (' . $list_keys . ')VALUES(' . $list_inter . ')', $values);
-    }
-
-    public function delete($id)
-    {
-        return $this->request("DELETE FROM comments WHERE id = ?", [$id]); 
-    }
-
-    public function hydrate($data): self
-    {
-        foreach ($data as $key => $value) {
-            // Retrieves setter corresponding to key
-            $setter = 'set' . ucfirst($key);
-            // Check if setter exists
-            if (method_exists($this, $setter)) {
-                // Call setter
-                $this->$setter($value);
-            }
-        }
-        return $this;
-    }
-
-    public function request(string $sql, array $params = null): PDOStatement | false
-    {
-        // Get instance of database
-        $this->db = Database::getInstance();
-
-        // Check if there are any parameters
-        if ($params !== null) {
-            // Prepared request
-            $query = $this->db->prepare($sql);
-            $query->execute($params);
-            return $query;
-        } else {
-            // Simple request
-            return $this->db->query($sql);
-        }
     }
 
 }
