+++ app/Views/pages/articles/index.html.twig
@@ -32,6 +32,23 @@
                     </div>                    
                 {% endfor %}
             </div>
+            <nav>
+                <ul class="pagination">
+                  {% if currentPage != 1 %}
+                  <li class="page-item">
+                    {% set previousPage = currentPage - 1 %}
+                    <a href="/articles?p={{ previousPage }}" class="page-link">Précédent</a>
+                  </li>    
+                  {% endif %}
+                  {% if lastPage != true %}
+                    <li class="page-item">
+                      {% set nextPage = currentPage + 1 %}
+                      <a href="/articles?p={{ nextPage }}" class="page-link">Suivant</a>
+                    </li>
+                  {% endif %}
+
+                </ul>
+              </nav>
         </div>
     </section>
 {% endblock %}
